SCP: -Wall warning reductions, Win32 console
sim_console.c: Streamline Win32 console output.

  std_output is set once, so there isn't any need to repeatedly invoke
  GetConsoleMode(). Add an output function pointer, std_output_writer,
  that points to WriteConsoleA for direct console output or WriteFile
  when output isn't to a Win32 console. Improves overall console output
  responsiveness on Win32.

  sim_os_poll_kbd_ready() is never called. Removed.

sim_sock.c, sim_sock.h:

  - Use inet_pton(), inet_ntop() when AF_INET6 is defined instead of
    inet_aton() and inet_ntoa().
  - sim_addr_acl_check: Rename inner bits declaration to acl_bits to
    avoid shadowing outer bits variable. Shadowing is generally considered
    poor style.
  - sim_write_sock: Return size_t.
  - sim_read_sock: Return ssize_t (because it needs to return -1.)

sim_tmxr.c, sim_tmxr.h:

  Fix size_t -> int type slicing/truncation warnings. All of the
  variables changed in struct tmln are array indices, which should be
  size_t. On many 64-bit platforms, ints (32-bit) are upconverted to
  size_t (64-bit), so this should produce slightly better compiled code

sim_printf_fmts.h:

  Rethink the SIZE_T_FMT format and use format macros that mimic C99's
  style -> SIM_PRIsize_t, SIM_PRIxsize_t, SIM_PRIXsize_t and
  SIM_PRIssize_t. Propagate into the affected simulators (format updates
  only.)

+++ sim_sock.h
@@ -127,15 +127,16 @@ SOCKET sim_connect_sock_ex (const char *sourcehostport, const char *hostport, co
 #define sim_connect_sock(hostport, default_host, default_port) sim_connect_sock_ex(NULL, hostport, default_host, default_port, SIM_SOCK_OPT_BLOCKING)
 SOCKET sim_accept_conn_ex (SOCKET master, char **connectaddr, int opt_flags);
 #define sim_accept_conn(master, connectaddr) sim_accept_conn_ex(master, connectaddr, 0)
-int sim_check_conn (SOCKET sock, int rd);
-int sim_read_sock (SOCKET sock, char *buf, int nbytes);
-int sim_write_sock (SOCKET sock, const char *msg, int nbytes);
-void sim_close_sock (SOCKET sock);
-const char *sim_get_err_sock (const char *emsg);
-SOCKET sim_err_sock (SOCKET sock, const char *emsg);
-int sim_getnames_sock (SOCKET sock, char **socknamebuf, char **peernamebuf);
-void sim_init_sock (void);
-void sim_cleanup_sock (void);
+
+int sim_check_conn(SOCKET sock, int rd);
+ssize_t sim_read_sock(SOCKET sock, char *buf, size_t nbytes);
+size_t sim_write_sock(SOCKET sock, const char *msg, size_t nbytes);
+void sim_close_sock(SOCKET sock);
+const char *sim_get_err_sock(const char *emsg);
+SOCKET sim_err_sock(SOCKET sock, const char *emsg);
+int sim_getnames_sock(SOCKET sock, char **socknamebuf, char **peernamebuf);
+void sim_init_sock(void);
+void sim_cleanup_sock(void);
 
 #ifdef  __cplusplus
 }
